Created separate test file for badgers pages (see code review comment)
